"ads2_m12_a1_1: 2/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ m12/ADS-2_Exam-2/ADS-2_Exam-2/Solution.java
@@ -16,7 +16,7 @@ public class Solution {
 			String[] arr = scan.nextLine().split(" ");
 			Edge edge = new Edge(Integer.parseInt(arr[0]),
                 Integer.parseInt(arr[1]),
-                Float.parseFloat(arr[2]));
+                Double.parseDouble(arr[2]));
             edgewgraph.addEdge(edge);
 
 }
